requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
python-dotenv>=1.0
numpy>=1.24
pandas>=2.0
Pillow>=10.0
requests>=2.31
bcrypt>=4.0
PyJWT>=2.8
